+++ src/provider/objectTreeItem.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
+import { utils } from 'httpyac';
 import { default as filesize } from 'filesize';
-import { inspect } from 'util';
 
 export interface ObjectItem {
   key: string;
@@ -30,7 +30,7 @@ export function toObjectItems(val: unknown): Array<ObjectItem> | undefined {
 export class ObjectTreeItem extends vscode.TreeItem {
   constructor(element: ObjectItem) {
     super(element.key);
-    this.tooltip = inspect(element.value);
+    this.tooltip = utils.stringifySafe(element.value, 2);
     const type = typeof element.value;
     this.description = type;
     this.command = {